# Fix selector-pseudo-element-no-unknown performance