--- a/.changeset/smooth-tigers-wink.md
+++ b/.changeset/smooth-tigers-wink.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `selector-pseudo-element-no-unknown` performance