--- a/lib/rules/selector-pseudo-element-no-unknown/index.js
+++ b/lib/rules/selector-pseudo-element-no-unknown/index.js
@@ -42,15 +42,15 @@ const rule = (primary, secondaryOptions) => {
 		}
 
 		root.walkRules((ruleNode) => {
-			if (!isStandardSyntaxRule(ruleNode)) {
-				return;
-			}
-
 			const selector = ruleNode.selector;
 
 			// Return early before parse if no pseudos for performance
 
 			if (!selector.includes(':')) {
+				return;
+			}
+
+			if (!isStandardSyntaxRule(ruleNode)) {
 				return;
 			}